pyproject.toml
[project]
name = "classifai"
description = "An AI-assisted flexible tool for different classification tasks."
authors = [
    { name = "Data Science Campus", email = "datasciencecampus@ons.gov.uk" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
    "pyprojroot==0.3.0",
    "fastapi[standard]>=0.111.0",
    "toml>=0.10.2",
    "astroid>=2.3.3",
    "chromadb>=0.6.0",
    "Click>=7.0",
    "colorama>=0.4.3",
    "fastapi>=0.111.0",
    "google-generativeai>=0.7.0",
    "google-genai>=1.8.0",
    "google-cloud-secret-manager>=2.20.2",
    "google-cloud-storage>=2.18.2",
    "google-cloud-logging>=3.11.2",
    "gunicorn>=20.0.4",
    "h11>=0.9.0",
    "isort>=4.3.21",
    "langchain-community>=0.2.7",
    "langchain-google-vertexai>=1.0.8",
    "lazy-object-proxy>=1.4.3",
    "matplotlib",
    "mccabe>=0.6.1",
    "pandas>=2.2.2",
    "polars>=1.0.0",
    "pydantic==2.9.2",
    "pylint>=2.4.4",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "seaborn",
    "scikit-learn",
    "six>=1.14.0",
    "starlette>=0.12.9",
    "sqlalchemy==2.0.35",
    "toml>=0.10.2",
    "uvicorn>=0.11.2",
    "websockets>=8.1",
    "wrapt>=1.11.2",
    "Flask>=3.0.3",
    "Flask-SQLAlchemy==3.1.1"
]


[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = [
    "hypothesis>=6.98.6",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-randomly>=3.15.0",
    "openpyxl>=3.1.5",
]
dev = [
    "pre-commit==3.3.3",
    "ruff==0.3.0",
    "classifai[test]",
    "ipykernel>=6.29.5",
    "mockito>=1.5.4",
]

[tool.ruff]
line-length = 79
exclude = ["uv.lock"]

[tool.ruff.lint]
extend-select = ["D", "I", "W"]
ignore = ["D105", "D107", "D202"]

[tool.ruff.lint.isort]
known-first-party = ["classifai"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
